refactor(llmobs): keep MCP dedup state private
